feat: move amaru ledger commands into amaru binary
