+++ crates/amaru/src/bin/amaru/cmd/dev/ledger/sync.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 PRAGMA
+// Copyright 2026 PRAGMA
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -36,7 +36,7 @@ use rayon::prelude::*;
 use tar::Archive;
 use tracing::info;
 
-use crate::cmd::new_block_validator;
+use super::new_block_validator;
 
 #[derive(Debug, clap::Parser)]
 pub struct Args {
@@ -183,7 +183,6 @@ async fn process_block(
 
     Ok(())
 }
-
 pub async fn run(args: Args) -> Result<(), Box<dyn std::error::Error>> {
     let network = args.network;
     let ledger_dir = args.ledger_dir.unwrap_or_else(|| default_ledger_dir(network).into());
