+++ scripts/refresh-from-mithril
@@ -183,13 +183,13 @@ sync_refresh() {
   local chain_dir="$1" ledger_dir="$2"
 
   echo "[mithril-refresh] downloading missing Mithril immutable chunks and packaging blocks"
-  cargo run --profile "$BUILD_PROFILE" --bin amaru-ledger -- mithril \
+  cargo run --profile "$BUILD_PROFILE" --bin amaru -- dev ledger mithril \
     --network "$NETWORK" \
     --ledger-dir "$ledger_dir" \
     --snapshots-dir "$SNAPSHOTS_DIR"
 
   echo "[mithril-refresh] syncing Amaru chain and ledger databases from packaged blocks"
-  cargo run --profile "$BUILD_PROFILE" --bin amaru-ledger -- sync \
+  cargo run --profile "$BUILD_PROFILE" --bin amaru -- dev ledger sync \
     --network "$NETWORK" \
     --chain-dir "$chain_dir" \
     --ledger-dir "$ledger_dir"
